feature/truncate clean up method

+++ src/main/java/com/github/vkorobkov/jfixtures/sql/SqlBase.java
@@ -1,5 +1,6 @@
 package com.github.vkorobkov.jfixtures.sql;
 
+import com.github.vkorobkov.jfixtures.config.structure.tables.CleanMethod;
 import com.github.vkorobkov.jfixtures.instructions.CleanTable;
 import com.github.vkorobkov.jfixtures.instructions.CustomSql;
 import com.github.vkorobkov.jfixtures.instructions.InsertRow;
@@ -19,17 +20,10 @@ public interface SqlBase extends Sql {
         val table = escapeTableOrColumn(cleanTable.getTable());
         val cleanMethod = cleanTable.getCleanMethod();
 
-        switch (cleanMethod) {
-        case DELETE:
+        if (cleanMethod == CleanMethod.DELETE) {
             appender.append("DELETE FROM ", table, ";\n");
-            break;
-        case TRUNCATE:
+        } else if (cleanMethod == CleanMethod.TRUNCATE) {
             appender.append("TRUNCATE TABLE ", table, ";\n");
-            break;
-        case NONE:
-            break;
-        default:
-            throw new IllegalStateException("Unknown clean method");
         }
     }
